+++ src/geometry/toi.rs
@@ -10,7 +10,7 @@ pub struct RawShapeColliderTOI {
 
 #[wasm_bindgen]
 impl RawShapeColliderTOI {
-    pub fn colliderHandle(&self) -> usize {
+    pub fn colliderHandle(&self) -> u32 {
         self.handle.into_raw_parts().0
     }
